src/absa_thesis/modeling/eval_mcnemar.py
"""McNemar test comparing RoBERTa predictions vs SVM baseline on the same test set.
This is a placeholder outline; actual integration requires persisted predictions.
"""
import numpy as np
from statsmodels.stats.contingency_tables import mcnemar
# This is a template. Save both model predictions to arrays y_pred_a, y_pred_b and ground truth y_true.
# Then build the 2x2 table on correctness and call mcnemar(table, exact=False, correction=True).
print("See comments in file for instructions to run McNemar once predictions are saved.")
