requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
python-dotenv==1.0.1
tqdm==4.66.4
loguru==0.7.2

# NLP + Modeling
scikit-learn==1.4.2
nltk==3.8.1
transformers==4.41.2
datasets==2.19.0
accelerate==0.30.1
torch>=2.2.0

# Rules
mlxtend==0.23.1

# IO + Viz
pyarrow==16.1.0
matplotlib==3.8.4
seaborn==0.13.2

# Dashboard
streamlit==1.36.0

# Testing
pytest==8.2.0

# Optional: Mongo
pymongo==4.7.2
